Fixes out of sequence event behavior
The pandora/event_support monitors Mopidy events and triggers pandora
events (thumbs up/down etc) that match certain sequences.  The default
action to trigger a thumbs_down is pause followed by next.  This
action produces the following sequence in testing:

1. track_playback_paused
2. track_playback_ended
3. track_playback_paused
4. track_changed_next

But in production (at least on my system) the sequence is:
1. track_playback_paused
2. track_playback_ended
3. track_changed_next
4. track_playback_paused

(notice the order of track_changed_next and track_playback_paused).

The relevant sequence is playback_paused, playback_ended,
track_changed_next.  The second playback_paused event is not relevant.
This intent is handled in the EventSequence class by way of the wait_for
option.  However, there is a small bug in the sequence matching that
causes the second pause to be matched, not ignored.

This commit fixes the issue so that the default event handling for
thumbs_down and sleep work as expected